Subtitles: Improve error checking with ASS rendering.
Some crashes could occur when the SSA script was using obviously wrong values.

Is edited to use an amount of samples that will either let one use
an effective maximum texture size of 8000x500 or 4000x1000.

This is more strict than the width >4000 OR height >4000 check
that was first decided on, but will let users create longer/higher
pieces of text that get moved on the screen in order to create an
illusion of scrolling.

Backport of MPC-HC commit 2c597d9